Move PackageWarningCounter into its own file
