+++ lib/src/warnings.dart
@@ -1,0 +1,259 @@
+import 'dart:io';
+
+import 'package:analyzer/dart/element/element.dart';
+import 'package:tuple/tuple.dart';
+
+import 'config.dart';
+
+class PackageWarningHelpText {
+  final String warningName;
+  final String shortHelp;
+  List<String> longHelp;
+  final PackageWarning warning;
+
+  PackageWarningHelpText(this.warning, this.warningName, this.shortHelp,
+      [this.longHelp]) {
+    this.longHelp ??= [];
+  }
+}
+
+/// Provides description text and command line flags for warnings.
+/// TODO(jcollins-g): Actually use this for command line flags.
+final Map<PackageWarning, PackageWarningHelpText> packageWarningText = {
+  PackageWarning.ambiguousDocReference: new PackageWarningHelpText(
+      PackageWarning.ambiguousDocReference,
+      "ambiguous-doc-reference",
+      "A comment reference could refer to two or more different objects"),
+  PackageWarning.ambiguousReexport: new PackageWarningHelpText(
+      PackageWarning.ambiguousReexport,
+      "ambiguous-reexport",
+      "A symbol is exported from private to public in more than one library and dartdoc can not determine which one is canonical",
+      [
+        "Use {@canonicalFor @@name@@} in the desired library's documentation to resolve",
+        "the ambiguity and/or override dartdoc's decision, or structure your package ",
+        "so the reexport is less ambiguous.  The symbol will still be referenced in ",
+        "all candidates -- this only controls the location where it will be written ",
+        "and which library will be displayed in navigation for the relevant pages.",
+        "The flag --ambiguous-reexport-scorer-min-confidence allows you to set the",
+        "threshold at which this warning will appear."
+      ]),
+  PackageWarning.ignoredCanonicalFor: new PackageWarningHelpText(
+      PackageWarning.ignoredCanonicalFor,
+      "ignored-canonical-for",
+      "A @canonicalFor tag refers to a library which this symbol can not be canonical for"),
+  PackageWarning.noCanonicalFound: new PackageWarningHelpText(
+      PackageWarning.noCanonicalFound,
+      "no-canonical-found",
+      "A symbol is part of the public interface for this package, but no library documented with this package documents it so dartdoc can not link to it"),
+  PackageWarning.noLibraryLevelDocs: new PackageWarningHelpText(
+      PackageWarning.noLibraryLevelDocs,
+      "no-library-level-docs",
+      "There are no library level docs for this library"),
+  PackageWarning.categoryOrderGivesMissingPackageName: new PackageWarningHelpText(
+      PackageWarning.categoryOrderGivesMissingPackageName,
+      "category-order-gives-missing-package-name",
+      "The category-order flag on the command line was given the name of a nonexistent package"),
+  PackageWarning.unresolvedDocReference: new PackageWarningHelpText(
+      PackageWarning.unresolvedDocReference,
+      "unresolved-doc-reference",
+      "A comment reference could not be found in parameters, enclosing class, enclosing library, or at the top level of any documented library with the package"),
+  PackageWarning.brokenLink: new PackageWarningHelpText(
+      PackageWarning.brokenLink,
+      "brokenLink",
+      "Dartdoc generated a link to a non-existent file"),
+  PackageWarning.orphanedFile: new PackageWarningHelpText(
+      PackageWarning.orphanedFile,
+      "orphanedFile",
+      "Dartdoc generated files that are unreachable from the index"),
+  PackageWarning.unknownFile: new PackageWarningHelpText(
+      PackageWarning.unknownFile,
+      "unknownFile",
+      "A leftover file exists in the tree that dartdoc did not write in this pass"),
+  PackageWarning.missingFromSearchIndex: new PackageWarningHelpText(
+      PackageWarning.missingFromSearchIndex,
+      "missingFromSearchIndex",
+      "A file generated by dartdoc is not present in the generated index.json"),
+  PackageWarning.typeAsHtml: new PackageWarningHelpText(
+      PackageWarning.typeAsHtml,
+      "typeAsHtml",
+      "Use of <> in a comment for type parameters is being treated as HTML by markdown"),
+};
+
+/// Something that package warnings can be called on.
+abstract class Warnable implements Locatable {
+  void warn(PackageWarning warning,
+      {String message, Iterable<Locatable> referredFrom});
+  Warnable get enclosingElement;
+}
+
+/// Something that can be located for warning purposes.
+abstract class Locatable {
+  String get fullyQualifiedName;
+  String get href;
+  List<Locatable> get documentationFrom;
+  Element get element;
+  String get elementLocation;
+  Tuple2<int, int> get lineAndColumn;
+  bool get isCanonical;
+}
+
+// The kinds of warnings that can be displayed when documenting a package.
+enum PackageWarning {
+  ambiguousDocReference,
+  ambiguousReexport,
+  ignoredCanonicalFor,
+  noCanonicalFound,
+  noLibraryLevelDocs,
+  categoryOrderGivesMissingPackageName,
+  unresolvedDocReference,
+  brokenLink,
+  orphanedFile,
+  unknownFile,
+  missingFromSearchIndex,
+  typeAsHtml,
+}
+
+class PackageWarningOptions {
+  // PackageWarnings must be in one of _ignoreWarnings or union(_asWarnings, _asErrors)
+  final Set<PackageWarning> _ignoreWarnings = new Set();
+  // PackageWarnings can be in both asWarnings and asErrors, latter takes precedence
+  final Set<PackageWarning> _asWarnings = new Set();
+  final Set<PackageWarning> _asErrors = new Set();
+
+  bool autoFlush = true;
+
+  Set<PackageWarning> get ignoreWarnings => _ignoreWarnings;
+  Set<PackageWarning> get asWarnings => _asWarnings;
+  Set<PackageWarning> get asErrors => _asErrors;
+
+  PackageWarningOptions() {
+    _asWarnings.addAll(PackageWarning.values);
+    ignore(PackageWarning.typeAsHtml);
+  }
+
+  void _assertInvariantsOk() {
+    assert(_asWarnings
+        .union(_asErrors)
+        .union(_ignoreWarnings)
+        .containsAll(PackageWarning.values.toSet()));
+    assert(_asWarnings.union(_asErrors).intersection(_ignoreWarnings).isEmpty);
+  }
+
+  void ignore(PackageWarning kind) {
+    _assertInvariantsOk();
+    _asWarnings.remove(kind);
+    _asErrors.remove(kind);
+    _ignoreWarnings.add(kind);
+    _assertInvariantsOk();
+  }
+
+  void warn(PackageWarning kind) {
+    _assertInvariantsOk();
+    _ignoreWarnings.remove(kind);
+    _asWarnings.add(kind);
+    _asErrors.remove(kind);
+    _assertInvariantsOk();
+  }
+
+  void error(PackageWarning kind) {
+    _assertInvariantsOk();
+    _ignoreWarnings.remove(kind);
+    _asWarnings.add(kind);
+    _asErrors.add(kind);
+    _assertInvariantsOk();
+  }
+}
+
+class PackageWarningCounter {
+  final _countedWarnings =
+      new Map<Element, Set<Tuple2<PackageWarning, String>>>();
+  final _warningCounts = new Map<PackageWarning, int>();
+  final PackageWarningOptions options;
+
+  final _buffer = new StringBuffer();
+
+  PackageWarningCounter(this.options);
+
+  /// Flush to stderr, but only if [options.autoFlush] is true.
+  ///
+  /// We keep a buffer because under certain conditions (--auto-include-dependencies)
+  /// warnings here might be duplicated across multiple Package constructions.
+  void maybeFlush() {
+    if (options.autoFlush) {
+      stderr.write(_buffer.toString());
+      _buffer.clear();
+    }
+  }
+
+  /// Actually write out the warning.  Assumes it is already counted with add.
+  void _writeWarning(PackageWarning kind, String name, String fullMessage) {
+    if (options.ignoreWarnings.contains(kind)) return;
+    String toWrite;
+    if (!options.asErrors.contains(kind)) {
+      if (options.asWarnings.contains(kind))
+        toWrite = "warning: ${fullMessage}";
+    } else {
+      if (options.asErrors.contains(kind)) toWrite = "error: ${fullMessage}";
+    }
+    if (toWrite != null) {
+      _buffer.write("\n ${toWrite}");
+      if (_warningCounts[kind] == 1 &&
+          config.verboseWarnings &&
+          packageWarningText[kind].longHelp.isNotEmpty) {
+        // First time we've seen this warning.  Give a little extra info.
+        final String separator = '\n            ';
+        final String nameSub = r'@@name@@';
+        String verboseOut =
+            '$separator${packageWarningText[kind].longHelp.join(separator)}';
+        verboseOut = verboseOut.replaceAll(nameSub, name);
+        _buffer.write(verboseOut);
+      }
+    }
+    maybeFlush();
+  }
+
+  /// Returns true if we've already warned for this.
+  bool hasWarning(Warnable element, PackageWarning kind, String message) {
+    Tuple2<PackageWarning, String> warningData = new Tuple2(kind, message);
+    if (_countedWarnings.containsKey(element?.element)) {
+      return _countedWarnings[element?.element].contains(warningData);
+    }
+    return false;
+  }
+
+  /// Adds the warning to the counter, and writes out the fullMessage string
+  /// if configured to do so.
+  void addWarning(Warnable element, PackageWarning kind, String message,
+      String fullMessage) {
+    assert(!hasWarning(element, kind, message));
+    Tuple2<PackageWarning, String> warningData = new Tuple2(kind, message);
+    _warningCounts.putIfAbsent(kind, () => 0);
+    _warningCounts[kind] += 1;
+    _countedWarnings.putIfAbsent(element?.element, () => new Set());
+    _countedWarnings[element?.element].add(warningData);
+    _writeWarning(kind, element?.fullyQualifiedName, fullMessage);
+  }
+
+  int get errorCount {
+    return _warningCounts.keys
+        .map((w) => options.asErrors.contains(w) ? _warningCounts[w] : 0)
+        .reduce((a, b) => a + b);
+  }
+
+  int get warningCount {
+    return _warningCounts.keys
+        .map((w) =>
+            options.asWarnings.contains(w) && !options.asErrors.contains(w)
+                ? _warningCounts[w]
+                : 0)
+        .reduce((a, b) => a + b);
+  }
+
+  @override
+  String toString() {
+    String errors = '$errorCount ${errorCount == 1 ? "error" : "errors"}';
+    String warnings =
+        '$warningCount ${warningCount == 1 ? "warning" : "warnings"}';
+    return [errors, warnings].join(', ');
+  }
+}
