+++ test/model_test.dart
@@ -10,6 +10,7 @@ import 'package:dartdoc/dartdoc.dart';
 import 'package:dartdoc/src/model.dart';
 import 'package:dartdoc/src/model_utils.dart';
 import 'package:dartdoc/src/package_meta.dart';
+import 'package:dartdoc/src/warnings.dart';
 import 'package:dartdoc/src/sdk.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
